Add Prometheus metrics support
Introduced Prometheus metrics integration for the application, enabling metrics reporting via a scheduled file writer when metrics are enabled. Updated dependencies to include necessary Prometheus libraries and modified configurations and documentation to support this new feature. The scheduling feature can be enabled by setting the environment variable METRICS_ENABLE to true. However, metrics will only be written to the /metrics endpoint if the vulnz CLI is running. Additionally, the container's version will be printed to the console during startup.