fix: preserve modified entries if year fails
If there is an exception on year 2017 and there were modified entries in the existing 2017 data file - the old data file would still be on disk for 2017 containing the modified entries - but when we write out the actual modified file, because 2017 was skipped we will not preserve the 2017 modified entries in the "modified" file. This PR corrects this.